t/tokenizer-test-cdata.dat

#data
<script><!-- </script> --></script>
#errors
unmatched </script>
#document-fragment p
| <script>
|   "<!-- "
| " -->"

#data
<script><!-- <script>abc</script> --></script>
#errors
#document-fragment p
| <script>
|   "<!-- <script>abc</script> -->"

#data
<style><!-- </style> --></style>
#errors
unmatched </style>
#document-fragment p
| <style>
|   "<!-- "
| " -->"

#data
<style><!-- <style>abc</style> --></style>
#errors
unmatched </style>
#document-fragment p
| <style>
|   "<!-- <style>abc"
| " -->"

#data
<script><!----x></script>foo
#errors
no DOCTYPE
#document
| <html>
|   <head>
|     <script>
|       "<!----x>"
|   <body>
|     "foo"

#data
<script><!-- <script></script> --x></script>foo
#errors
no DOCTYPE
#document
| <html>
|   <head>
|     <script>
|       "<!-- <script></script> --x>"
|   <body>
|     "foo"
